+++ test/cli/commands/database-init.spec.js
@@ -3,8 +3,8 @@ const cli = require('../../../src/cli');
 
 let mongod;
 beforeAll(async () => {
-    mongod = new MongoMemoryServer();
-    process.env.MONGODB_URI = await mongod.getUri();
+    mongod = await MongoMemoryServer.create();
+    process.env.MONGODB_URI = mongod.getUri();
 });
 afterAll(async () => {
     await mongod.stop();