+++ test/server/api/files.spec.js
@@ -12,8 +12,8 @@ process.env.NODE_CONFIG = JSON.stringify({ uploads: { path: uploadPath } });
 let app;
 let mongod;
 beforeAll(async () => {
-    mongod = new MongoMemoryServer();
-    process.env.MONGODB_URI = await mongod.getUri();
+    mongod = await MongoMemoryServer.create();
+    process.env.MONGODB_URI = mongod.getUri();
     app = require('../../../server/app');
 });
 afterAll(async () => {