+++ test/server/api/version.spec.js
@@ -5,8 +5,8 @@ const { MongoMemoryServer } = require('mongodb-memory-server');
 let app;
 let mongod;
 beforeAll(async () => {
-    mongod = new MongoMemoryServer();
-    process.env.MONGODB_URI = await mongod.getUri();
+    mongod = await MongoMemoryServer.create();
+    process.env.MONGODB_URI = mongod.getUri();
     app = require('../../../server/app');
 });
 afterAll(async () => {