+++ test/server/helpers/mongoose.spec.js
@@ -3,8 +3,8 @@ const mongoose = require('mongoose');
 
 let mongod;
 beforeAll(async () => {
-    mongod = new MongoMemoryServer();
-    process.env.MONGODB_URI = await mongod.getUri();
+    mongod = await MongoMemoryServer.create();
+    process.env.MONGODB_URI = mongod.getUri();
     require('../../../server/app');
 });
 afterAll(async () => {