+++ test/server/setup.js
@@ -2,7 +2,6 @@ const { MongoMemoryServer } = require('mongodb-memory-server');
 
 module.exports = async () => {
     // Make sure the mongod binary is downloaded before running tests
-    const mongod = new MongoMemoryServer();
-    await mongod.getUri();
+    const mongod = await MongoMemoryServer.create();
     await mongod.stop();
 };